pyproject.toml
[project]
name = "vehicledetector"
version = "0.1.0"
description = ""
authors = [
    {name = "aeSynk",email = "daniyar211998@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ultralytics (>=8.3.129,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
